# docs/specs/architecture/openapi/social.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "xitter social API",
    "description": "Profiles, follows, and blocks.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/api/social/v1"
    }
  ],
...
  "paths": {
...
    "/profiles/{userId}/follow": {
      "post": {
        "tags": ["relationships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Following"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "tags": ["relationships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Unfollowed"
          }
        }
      }
    },
    "/profiles/{userId}/block": {
      "post": {
        "tags": ["relationships"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Blocked"
          }
        }
      }
    },
...
  }
}